requirements.txt
streamlit
pandas
numpy
scikit-learn
joblib
matplotlib
seaborn
google-generativeai
python-dotenv
plotly
pyttsx3
